+++ packages/editor/js/extensions/libs/background/test/background-color.e2e.cy.js
@@ -102,7 +102,7 @@ describe('Background Color → Functionality', () => {
 		//assert frontend
 		savePage();
 		redirectToFrontPage();
-		cy.get('style#blockera-inline-css-inline-css')
+		cy.get('style#blockera-inline-css')
 			.invoke('text')
 			.should(
 				'include',