+++ packages/editor/js/extensions/libs/effects/test/filters.e2e.cy.js
@@ -87,7 +87,7 @@ describe('Filters → Functionality', () => {
 
 		redirectToFrontPage();
 
-		cy.get('style#blockera-inline-css-inline-css')
+		cy.get('style#blockera-inline-css')
 			.invoke('text')
 			.should(
 				'include',