+++ packages/editor/js/extensions/libs/size/test/min-width.e2e.cy.js
@@ -89,7 +89,7 @@ describe('Min Width → Functionality', () => {
 
 		redirectToFrontPage();
 
-		cy.get('style#blockera-inline-css-inline-css')
+		cy.get('style#blockera-inline-css')
 			.invoke('text')
 			.should(
 				'include',