+++ packages/editor/js/extensions/libs/typography/test/font-color.e2e.cy.js
@@ -89,7 +89,7 @@ describe('Font Color → Functionality', () => {
 
 		redirectToFrontPage();
 
-		cy.get('style#blockera-inline-css-inline-css')
+		cy.get('style#blockera-inline-css')
 			.invoke('text')
 			.should(
 				'include',